#105 - move the dialects json to a database folder
