+++ controllers/dialects/find.js
@@ -1,4 +1,4 @@
-const dialects = require('../../dialects')
+const dialects = require('../../database/dialects')
 
 module.exports = (req, res) => {
   const { region } = req.params
