+++ controllers/dialects/get.js
@@ -1,4 +1,4 @@
-const dialects = require('../../dialects')
+const dialects = require('../../database/dialects')
 
 module.exports = (req, res) => {
   const dialect = req.query.dialetus
